refactor(editor): extract vertical alignment post-processing
Move the 84-line #applyVertAlignToLayout method into a pure function
that takes a PM doc and painterHost. Handles subscript/superscript
CSS styling by scanning painted spans, resolving run properties,
and applying vertical-align and font-size styles.

Also removes unused imports (decodeRPrFromMarks, halfPointToPoints)
and constant (SUBSCRIPT_SUPERSCRIPT_SCALE) from PresentationEditor.

Reduces PresentationEditor from 5167 to 5082 lines (-85).
